+++ resources/lang/de/bank.php
@@ -6,4 +6,6 @@ return [
     'view_bank_report' => 'Bankbericht anzeigen',
     'frequent_visitors_explanation' => 'Personen werden als regelmässige Besucher markiert (\'gelbes Stern\' Symbol) wenn sie mindestens _x_ mal während der letzten _y_ Wochen vorbeikommen.',
     'frequent_visitors_affected' => 'Aktuelle Einstellungen: :freq Personen betroffen, von total :total (:percentage %)',
+    'label_on_code_card' => 'Text auf Codekarte',
+    'logo_on_code_card' => 'Logo auf Codekarte',
 ];