+++ resources/lang/en/bank.php
@@ -6,4 +6,6 @@ return [
     'view_bank_report' => 'View bank report',
     'frequent_visitors_explanation' => 'Persons are marked as frequent visitor (\'yellow star\' icon) if they visit the bank at least _x_ times during the last _y_ weeks.',
     'frequent_visitors_affected' => 'Current settings: :freq persons affected, out of :total (:percentage %)',
+    'label_on_code_card' => 'Label on code card',
+    'logo_on_code_card' => 'Logo on code card',
 ];