+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## NEXT
+
+* Added support for file / image upload in settings editor
+* Bank code card image can now be defined in settings view
+
 ## 2.12.1
 
 * Optimized bank statistics database query